refactor!: remove unused components 3/3 (#1483)
* refactor(header): move header component into globalheader recipe

* refactor(primarynav): move subcomponent within component file

* refactor(primarynav): move to recipe

* refactor: shorthand prop={true} (#1484)

* refactor: shorthand prop={true}

* refactor(lint): turn on jsx boolean value lint
