+++ .storybook/recipes/PrimaryNav/PrimaryNav.module.css
@@ -1,4 +1,4 @@
-@import '../../design-tokens/mixins.css';
+@import '../../../src/design-tokens/mixins.css';
 
 /*------------------------------------*\
     # PRIMARY NAVIGATION ITEM
@@ -41,17 +41,16 @@
       @mixin focusInverted;
     }
   }
-
-  /**
-  * Primary nav link within active primary nav item
-  */
-  .primary-nav__item.eds-is-active & {
-    background: var(--eds-theme-color-button-primary-brand-background-active);
-  }
-
   @media screen and (prefers-reduced-motion) {
     transition: none;
   }
+}
+
+/**
+ * Active primary nav link
+ */
+.primary-nav__link--active {
+  background: var(--eds-theme-color-button-primary-brand-background-active);
 }
 
 /**
@@ -60,4 +59,33 @@
 .primary-nav__icon {
   margin-right: var(--eds-size-2-and-half);
   color: var(--eds-theme-color-text-neutral-default-inverse);
+}
+
+/*------------------------------------*\
+    # PRIMARY NAVIGATION
+\*------------------------------------*/
+
+/**
+ * Primary navigation existing in the header and maybe the footer.
+ * 
+ * The outer <nav> wrapper of the primary navigation.
+ */
+
+/**
+ * Primary nav list
+ * 
+ The actual list of nav items.
+ */
+.primary-nav__list {
+  @mixin reset;
+  list-style: none;
+  display: flex;
+  flex-direction: column;
+  gap: var(--eds-size-1);
+  width: fit-content;
+  z-index: var(--eds-z-index-200);
+
+  @media all and (min-width: $eds-bp-xl) {
+    width: auto;
+  }
 }
