+++ docs/CODE_GUIDELINES.md
@@ -574,7 +574,7 @@ The default option should be the one most commonly used in order to reduce frict
 - `size` should be used for adjusting size attributes (e.g. `<Button variant="secondary" size="sm">` or `<Button size="md"`>). Use abbreviations for sizes (ex: `xs`, `sm`, `md`, `lg`).
 - `behavior` should be used for functional variations of a pattern, such as `<Banner behavior="dismissable">`. Additional non-exclusive behaviors should be handled using boolean props prefixed with `is` (e.g. `isSticky` and `isDismissable`).
 - `orientation` should be used for controlling the layout or orientation of a component (e.g. `<ButtonGroup orientation="stacked">`)
-- `disabled` boolean should be used to control the interactivity of a component (e.g. `<Button disabled={true} />`)
+- `disabled` boolean should be used to control the interactivity of a component (e.g. `<Button disabled />`)
 - `align` should be used for aligning content, and should include `left` (default), `center`, `right` if needed.
 - `verticalAlign` should be used for vertically aligning content, and should include `top`, `middle`, `bottom` if needed.
