--- src/components/Header/Header.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as stories from './Header.stories';
-
-describe('<Header />', () => {
-  generateSnapshots(stories);
-});
