+++ src/components/Modal/Modal.stories.tsx
@@ -277,7 +277,7 @@ export const ControlHeadingInteractive: StoryObj<HeadingArgs> = {
 
 export const WithoutCloseButton: StoryObj<InteractiveArgs> = {
   render: (args) => (
-    <InteractiveExample {...args} hideCloseButton={true}>
+    <InteractiveExample {...args} hideCloseButton>
       {getChildren()}
     </InteractiveExample>
   ),
