+++ src/components/Modal/Modal.test.tsx
@@ -76,7 +76,7 @@ describe('Modal', () => {
 
   it('does not throw an error if modal uses <Modal.Title>', () => {
     const modalWithTitle = (
-      <Modal onClose={() => {}} open={true}>
+      <Modal onClose={() => {}} open>
         <Modal.Header>
           <Modal.Title>Modal Title</Modal.Title>
         </Modal.Header>
@@ -93,7 +93,7 @@ describe('Modal', () => {
 
   it('does not throw an error if modal uses aria-label', () => {
     const modalWithAriaLabel = (
-      <Modal aria-label="aria label" onClose={() => {}} open={true}>
+      <Modal aria-label="aria label" onClose={() => {}} open>
         <Modal.Header>Modal Title</Modal.Header>
         <Modal.Body>Modal body content.</Modal.Body>
         <Modal.Footer>Modal footer content.</Modal.Footer>
@@ -108,7 +108,7 @@ describe('Modal', () => {
 
   it('does throw an error if modal does not use <Modal.Title> or aria-label', () => {
     const modalWithoutTitleOrAriaLabel = (
-      <Modal onClose={() => {}} open={true}>
+      <Modal onClose={() => {}} open>
         <Modal.Header>Modal Title</Modal.Header>
         <Modal.Body>Modal body content.</Modal.Body>
         <Modal.Footer>Modal footer content.</Modal.Footer>
