--- src/components/PrimaryNav/PrimaryNav.module.css
@@ -1,30 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # PRIMARY NAVIGATION
-\*------------------------------------*/
-
-/**
- * Primary navigation existing in the header and maybe the footer.
- * 
- * The outer <nav> wrapper of the primary navigation.
- */
-
-/**
- * Primary nav list
- * 
- The actual list of nav items.
- */
-.primary-nav__list {
-  @mixin reset;
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: var(--eds-size-1);
-  width: fit-content;
-  z-index: var(--eds-z-index-200);
-
-  @media all and (min-width: $eds-bp-xl) {
-    width: auto;
-  }
-}
