--- src/components/PrimaryNav/PrimaryNav.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as stories from './PrimaryNav.stories';
-
-describe('<PrimaryNav />', () => {
-  generateSnapshots(stories);
-});
